Stable

+++ be/src/store/store.controller.ts
@@ -1,18 +1,18 @@
-import { Body, Controller, Delete, Get, Param, Post, Put, Req, UseGuards } from '@nestjs/common';
+import { Body, Controller, Delete, Get, Param, Post, Put, Query, Req, UploadedFile, UseGuards, UseInterceptors } from '@nestjs/common';
 import { StoreService } from './store.service';
-import { Store } from './schema/store.schema';
-import { Request } from 'express';
-import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';
-import { AbilitiesGuard } from 'src/ability/guards/abilities.guard';
-import { CheckAbilities, CreateBillAbility, CreateStoreAbility, DeleteStoreAbility, ReadStoreAbility, UpdateStoreAbility } from 'src/ability/decorators/abilities.decorator';
-import { Types } from 'mongoose';
+import { ApiBearerAuth, ApiConsumes, ApiCreatedResponse, ApiTags } from '@nestjs/swagger';
+import { AbilitiesGuard } from '../ability/guards/abilities.guard';
+import { CheckAbilities, CreateBillAbility, CreateStoreAbility, DeleteStoreAbility, ReadStoreAbility, UpdateStoreAbility } from '../ability/decorators/abilities.decorator';
 import { CreateStoreDto } from './dto/create-store.dto';
-import { UserService } from 'src/user/user.service';
-import { RoleService } from 'src/role/role.service';
-import { RoleName } from 'src/role/schema/role.schema';
-import { CheckRole } from 'src/ability/decorators/role.decorator';
-import { GetCurrentUserId } from 'src/auth/decorators/get-current-userid.decorator';
+import { UserService } from '../user/user.service';
+import { RoleService } from '../role/role.service';
+import { RoleName } from '../role/schema/role.schema';
+import { CheckRole } from '../ability/decorators/role.decorator';
+import { GetCurrentUserId } from '../auth/decorators/get-current-userid.decorator';
 import { UpdateStoreDto } from './dto/update-store.dto';
+import { BadRequestException, ConflicException, NotFoundException } from '../core/error.response';
+import { SuccessResponse } from '../core/success.response';
+
 
 @Controller('store')
 @ApiTags('Store')
@@ -31,11 +31,23 @@ export class StoreController {
   async create(
     @Body() store: CreateStoreDto,
     @GetCurrentUserId() userId: string,
-  ): Promise<Store> {
+  ): Promise<SuccessResponse | NotFoundException | ConflicException | BadRequestException> {
     const user = await this.userService.getById(userId)
-    const newStore = await this.storeService.create(user, store)
-    await this.roleService.addUserToRole(userId, { name: RoleName.SELLER })
-    return newStore
+    if (!user) return new NotFoundException("Không tìm thấy người dùng này!")
+
+    const hasStore = await this.storeService.getByUserId(userId)
+    if (hasStore) return new ConflicException("Người dùng này đã có cửa hàng!")
+
+    const newStore = await this.storeService.create(userId, store)
+    if (!newStore) return new BadRequestException("Tạo cửa hàng thất bại!")
+
+    const resultAddRole = await this.roleService.addUserToRole(userId, { name: RoleName.SELLER })
+    if (!resultAddRole) return new BadRequestException("Thêm quyền thất bại!")
+
+    return new SuccessResponse({
+      message: "Tạo cửa hàng thành công!",
+      metadata: { data: newStore },
+    })
   }
 
   @UseGuards(AbilitiesGuard)
@@ -44,20 +56,28 @@ export class StoreController {
   @Get('user/:id')
   async getById(
     @Param('id') id: string
-  ): Promise<Store> {
+  ): Promise<SuccessResponse | NotFoundException> {
     const store = await this.storeService.getById(id)
-    return store
+    if (!store) return new NotFoundException("Không tìm thấy cửa hàng này!")
+    return new SuccessResponse({
+      message: "Lấy thông tin cửa hàng thành công!",
+      metadata: { data: store },
+    })
   }
 
   @UseGuards(AbilitiesGuard)
   @CheckAbilities(new ReadStoreAbility())
-  @CheckRole(RoleName.SELLER)
+  @CheckRole(RoleName.SELLER, RoleName.USER)
   @Get('seller')
   async getMyStore(
     @GetCurrentUserId() userId: string,
-  ): Promise<Store> {
+  ): Promise<SuccessResponse | NotFoundException> {
     const store = await this.storeService.getByUserId(userId)
-    return store
+    if (!store) return new NotFoundException("Không tìm thấy cửa hàng này!")
+    return new SuccessResponse({
+      message: "Lấy thông tin cửa hàng thành công!",
+      metadata: { data: store },
+    })
   }
 
   @UseGuards(AbilitiesGuard)
@@ -67,9 +87,13 @@ export class StoreController {
   async update(
     @Body() store: UpdateStoreDto,
     @GetCurrentUserId() userId: string,
-  ): Promise<Store> {
+  ): Promise<SuccessResponse | NotFoundException> {
     const newStore = await this.storeService.update(userId, store)
-    return newStore
+    if (!newStore) return new NotFoundException("Không tìm thấy cửa hàng này!")
+    return new SuccessResponse({
+      message: "Cập nhật thông tin cửa hàng thành công!",
+      metadata: { data: newStore },
+    })
   }
 
 
@@ -79,20 +103,30 @@ export class StoreController {
   @Delete('seller')
   async delete(
     @GetCurrentUserId() userId: string,
-  ): Promise<boolean> {
-    await this.storeService.delete(userId)
+  ): Promise<SuccessResponse | NotFoundException | BadRequestException> {
+    const result = await this.storeService.delete(userId)
+    if (!result) return new NotFoundException("Không tìm thấy cửa hàng này!")
     const isDeleted = await this.roleService.removeUserRole(userId, RoleName.SELLER)
-    return isDeleted
+    if (!isDeleted) return new BadRequestException("Xóa quyền thất bại!")
+    return new SuccessResponse({
+      message: "Xóa cửa hàng thành công!",
+      metadata: { data: result },
+    })
   }
 
   @UseGuards(AbilitiesGuard)
   @CheckAbilities(new UpdateStoreAbility())
   @CheckRole(RoleName.MANAGER)
   @Put('manager/warningcount/:id')
-  async updateWarningCount(@Param('id') id: string, @Param("action") action: string): Promise<Store> {
+  async updateWarningCount(@Param('id') id: string, @Param("action") action: string): Promise<SuccessResponse | NotFoundException> {
     const store = await this.storeService.updateWarningCount(id, action);
-    return store
+    if (!store) throw new NotFoundException("Không tìm thấy cửa hàng này!")
+    return new SuccessResponse({
+      message: "Cập nhật cảnh báo thành công!",
+      metadata: { data: store },
+    })
   }
+
 }
